Make a grid movie and skip the step if it is already done.

+++ make-movie.sh
@@ -1,24 +1,27 @@
 #!/bin/bash
-rm -rf tmp
-mkdir tmp
-
 FPS=10
-
 i=1
 
-for TIF in $(ls *tif)
-do
-  DATA_NAME=${TIF/.tif/}
-  TMP=$(printf "tmp/image%03d.bmp" $i)
-  echo $TIF "-->" $TMP
-  convert $TIF label:\"$DATA_NAME\" -gravity Center -append $TMP
-  i=$(($i + 1))
-done
-
-
-DATA_NAME=$(ls $TIF | awk -F"_" '{print $1}')
-WELL_ADDRESS=$(ls $TIF | awk -F"." '{print $2}')
+DATA_NAME=$(basename $(dirname $PWD))
+WELL_ADDRESS=$(basename $PWD)
 MOVIE_NAME=$DATA_NAME"."$WELL_ADDRESS".avi"
-mencoder "mf://tmp/*.bmp" -mf fps=$FPS -o $MOVIE_NAME -ovc lavc
 
-rm -rf tmp
+if [ -e $MOVIE_NAME ]; then
+  echo "$MOVIE_NAME exists. Skip."
+else
+  rm -rf tmp
+  mkdir tmp
+
+  for TIF in $(ls *.{tif,jpg})
+  do
+    TMP=$(printf "tmp/image%03d.bmp" $i)
+    echo $TIF "-->" $TMP
+    convert $TIF label:\"$DATA_NAME\" -gravity Center -append $TMP
+    i=$(($i + 1))
+  done
+
+  #DATA_NAME=$(ls $TIF | awk -F"_" '{print $1}')
+  #WELL_ADDRESS=$(ls $TIF | awk -F"." '{print $2}')
+  mencoder "mf://tmp/*.bmp" -mf fps=$FPS -o $MOVIE_NAME -ovc lavc
+  rm -rf tmp
+fi
